requirements.txt
pandas
python-dotenv
supabase
resend
requests
openai